requirements.txt
flask==3.0.0
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
gunicorn==21.2.0
